+++ src/Directory.Build.props
@@ -1,7 +1,7 @@
 <Project>
   <PropertyGroup>
     <LangVersion>latest</LangVersion>
-    <ShaderGenVersion>1.2.100-beta2-gfab524daa3</ShaderGenVersion>
-    <VeldridVersion>4.3.3</VeldridVersion>
+    <ShaderGenVersion>1.2.106-beta2-gb17d8a8a99</ShaderGenVersion>
+    <VeldridVersion>4.4.0-beta3-gca5ebdedbf</VeldridVersion>
   </PropertyGroup>
 </Project>
